requirements.txt
#######raw_data/


#Packages to install
pillow
matplotlib
numpy
face_recognition
